requirements.txt
opencv-python-headless==4.8.1.78
ultralytics
numpy==1.26.4
pytube
PyYAML
torch==2.1.2
torchvision==0.16
requests
pillow
yt-dlp
torchvision
matplotlib
opencv-python
tqdm
scipy
seaborn
sympy
transformers
scikit-learn
pandas